.pipeline: Start integration with pipelinelib
This is a work in progress to add configuration to integrate with the
Foundation's <https://wikitech.wikimedia.org/wiki/PipelineLib> project.
The first step being taken here is describing a replacement for the
hand-built Dockerfile currently found in the root of the project with
a .pipeline/blubber.yaml configuration file for Blubber.