Makefile: add make-admin-user target
Add a Makefile target that will create a local superuser account named
'admin' if one does not already exist. The password for this account is
expected to be sent in the DJANGO_SUPERUSER_PASSWORD environment
variable. make_env.sh will generate a random password that can be found
in the generated .env configuration file.